@@ -14,18 +14,17 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=3.0.0 <3.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.13.6
   intl: ^0.18.1
   cookie_jar: ^4.0.3
   dio: ^5.1.2
   dio_cookie_manager: ^3.0.0
   path_provider: ^2.0.15
-  flutter_html: ^2.1.0
+  flutter_html: ^3.0.0-beta.2
 
 dev_dependencies:
   flutter_test:
