@@ -1,4 +1,4 @@
-package com.example.worldbooks
+package com.example.wb_flutter
 
 import io.flutter.embedding.android.FlutterActivity
 
